packages/duecare-llm-evidence-db/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "duecare-llm-evidence-db"
version = "0.1.0"
description = "Duecare evidence database: persists pipeline outputs (entities, edges, findings, tool-call results) to a queryable store. Local DuckDB by default; SQLite and PostgreSQL backends available."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "Taylor Amarel"}]
keywords = ["llm", "safety", "gemma", "duecare", "evidence", "duckdb"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Database",
]
dependencies = [
    "pydantic>=2.9.0",
    "duckdb>=1.0.0",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9.0"]
all = ["psycopg2-binary>=2.9.0"]

[project.urls]
Homepage = "https://github.com/TaylorAmarelTech/gemma4_comp"
Repository = "https://github.com/TaylorAmarelTech/gemma4_comp"

[tool.hatch.build.targets.wheel]
packages = ["src/duecare"]
